v1/requirements.txt
streamlit>=1.28.0
crewai>=0.20.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.0.0 